Reword questions and question subheadings
Add "The" to the first question's subheading.  Change the punctuation
mark at the end of the first question from a question mark to a period.

Change the second question's ending colon to a period.

Change the third subheading from "Asynchronous Code" to "Async
Advantages and Disadvantages".  Reword the third question and replace
"benefits" with "advantages".